Move PackageConfig related tests to test_package_config.py
